Deleting GitHub conflict syntax
I am keeping what I believe to be the correct code that we wrote for
this file. If you get a conflict in your file, please resolve them
before commiting the changes. This happens when you pull from GitHub and
there are changes that can't be combined with your local changes. Just
choose one section to be the final one, and get the code compiling.