API 27 Update
Making an update to all modules (TeamCode, FTCRobotController, navx_ftc)
to use API 27 and build tools 27.0.1. This is to be fully compatible
with the new gradle version (we used to be on API 19, which is now
several years old). Please download API 27 and build tools 27.0.1 (you
should be prompted to).

I also deleted some unnecessary depency lines from some of the gradle
build files. Maybe this will solve some people's issues.